Make `TagContext.submissionsData` ignore a trailing "[]"
Motivation: Vapor's `URLEncodedFormDecoder` expects array-typed fields to end with a `[]`. For example, to decode

```
struct Args: Codable {
  let options: [String]
}
```

one would need to pass in a query string/form of the form `options[]=foo&options[]=bar`. To achieve that, the input field's name in the Leaf template needs to be `options[]`, however, the Field's name is derived from its key path, which is `options`. To make this `submissionsData()` find the correct field, we also attempt searching for the stripped field path.